clean up code by extracting fraction and rational modules
