+++ Cargo.toml
@@ -17,15 +17,14 @@ codec = { package = "parity-scale-codec", version = "3.0.0", default-features = 
 sp-arithmetic = { git = "https://github.com/paritytech/substrate", branch="polkadot-v0.9.29", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate", branch="polkadot-v0.9.29", default-features = false }
 
-rug = { version = "1.17.0", features = ["num-traits"] }
 
 [dev-dependencies]
+approx = "0.5.1"
+criterion = "0.4.0"
+proptest = "1.0.0"
 rand = { version = "0.8.5", default-features = false, features= ["std_rng"] }
 rand_xoshiro = "0.6.0"
-criterion = "0.4.0"
-approx = "0.5.1"
-proptest = "1.0.0"
-
+rug = { version = "1.17.0", features = ["num-traits"] }
 test-case = "2.2.1"
 
 [features]
