+++ src/lib.rs
@@ -17,9 +17,11 @@ mod test_pow_accuracy;
 
 pub mod ema;
 pub mod fee;
+pub mod fraction;
 pub mod lbp;
 pub mod liquidity_mining;
 pub mod omnipool;
+pub mod rational;
 pub mod stableswap;
 #[cfg(test)]
 pub mod test_utils;
