+++ src/transcendental.rs
@@ -282,7 +282,8 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::types::{fraction, FixedBalance, Fraction};
+    use crate::fraction;
+    use crate::types::{FixedBalance, Fraction};
     use core::str::FromStr;
     use fixed::traits::LossyInto;
     use fixed::types::U64F64;
